annotation tool/SharkAnnotationStreamlit/requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
